Show both hands
Note the dealer shows their hole card. Also note this code is a dumpster
fire filled with poop, and desperately needs refactoring. BUT FORWARD WE
SHALL GO!

Right now the dealer hand and player hand are scenes. I am not sure this
is necessary, but I suspect it will help with animations later.